added upload image functionality to edit campgrounds

+++ routes/campgrounds.js
@@ -4,13 +4,16 @@ const Campground = require("../models/campground");
 const catchAsync = require("../utils/catchAsync");
 const Joi = require("joi");
 const { isLoggedin } = require("../middlewares");
+const multer = require("multer");
+const { storage } = require("../cloudinary");
+const upload = multer({ storage });
 
 // validating Campgrounds Creation
 const validateCampgrounds = (req, res, next) => {
   const campgroundSchema = Joi.object({
     title: Joi.string().required(),
     price: Joi.number().required().min(0),
-    image: Joi.string().required(),
+    // image: Joi.string().required(),
     description: Joi.string().required(),
     location: Joi.string().required(),
   }).required();
@@ -39,11 +42,15 @@ router.get("/new", isLoggedin, (req, res) => {
 router.post(
   "/",
   isLoggedin,
+  upload.array("image"),
   validateCampgrounds,
   catchAsync(async (req, res, next) => {
     try {
       const newCampground = new Campground(req.body);
+      newCampground.image = req.files.map((file) => ({ url: file.path, filename: file.filename }));
+      newCampground.author = req.user._id;
       await newCampground.save();
+      console.log(newCampground);
       req.flash("success", "Successfully made a new Campground!");
       res.redirect("/campgrounds");
     } catch (e) {
@@ -59,7 +66,9 @@ router.get(
   catchAsync(async (req, res, next) => {
     try {
       const { id } = req.params;
-      const campground = await Campground.findById(id).populate("reviews");
+      const campground = await Campground.findById(id)
+        .populate({ path: "reviews", populate: { path: "author" } })
+        .populate("author");
       if (!campground) {
         req.flash("error", "No Campground was found");
         return res.redirect("/campgrounds");
@@ -89,15 +98,27 @@ router.get(
 router.put(
   "/:id",
   isLoggedin,
-  validateCampgrounds,
+  upload.array("image"),
   catchAsync(async (req, res, next) => {
     try {
       const { id } = req.params;
-      const campground = await Campground.findByIdAndUpdate(id, {
-        ...req.body,
-      });
-      req.flash("success", "Successfully edited a Campground!");
-      res.redirect(`/campgrounds/${campground._id}`);
+      const campground = await Campground.findById(id);
+      const imgs = req.files.map((f) => ({ url: f.path, filename: f.filename }));
+      if (!campground.author.equals(req.user._id)) {
+        req.flash("error", "Not authorized to edit this campground");
+        res.redirect(`/campgrounds/${id}`);
+      } else {
+        campground.title = req.body.title;
+        campground.price = req.body.price;
+        campground.description = req.body.description;
+        campground.location = req.body.location;
+        campground.image.push(...imgs);
+
+        await campground.save();
+
+        req.flash("success", "Successfully edited a Campground!");
+        res.redirect(`/campgrounds/${campground._id}`);
+      }
     } catch (e) {
       next(e);
     }
@@ -108,9 +129,15 @@ router.put(
 router.delete("/:id", isLoggedin, async (req, res, next) => {
   try {
     const { id } = req.params;
-    await Campground.findByIdAndDelete(id);
-    req.flash("success", "Successfully deleted a Campground!");
-    res.redirect("/campgrounds");
+    const campgroundId = await Campground.findById(id);
+    if (!campgroundId.author.equals(req.user._id)) {
+      req.flash("error", "not authorized to edit this campground");
+      res.redirect(`/campgrounds/${id}`);
+    } else {
+      await Campground.findByIdAndDelete(id);
+      req.flash("success", "Successfully deleted a Campground!");
+      res.redirect("/campgrounds");
+    }
   } catch (e) {
     next(e);
   }
